+++ src/minotaur_pc/src/MapCreator.cpp
@@ -20,102 +20,102 @@
 using namespace std;
 
 struct sensorDistanceFromZero {
-  int x;
-  int y;
+    int x;
+    int y;
 } sen1, sen2, sen3;
 
 struct position {
-  float x;
-  float y;
-  float theta;
+    float x;
+    float y;
+    float theta;
 } pos;
 
 namespace minotaur {
-  
+    
     // Default Constructor
     MapCreator::MapCreator()
     {
-      map.setDimension(500, 500);
-      
-      // set Default-values for sensors:
-      setSensorDistances(1, 0, 0);
-      setSensorDistances(2, 0, 0);
-      setSensorDistances(3, 0, 0);
+        map.setDimension(500, 500);
+        
+        // set Default-values for sensors:
+        setSensorDistances(1, 0, 0);
+        setSensorDistances(2, 0, 0);
+        setSensorDistances(3, 0, 0);
     }
-  
+    
     MapCreator::MapCreator(int width, int height)
     {
-      map.setDimension(width, height);
-      
-      // set Default-values for sensors:
-      setSensorDistances(1, 0, 0);
-      setSensorDistances(2, 0, 0);
-      setSensorDistances(3, 0, 0);
-      
+        map.setDimension(width, height);
+        
+        // set Default-values for sensors:
+        setSensorDistances(1, 0, 0);
+        setSensorDistances(2, 0, 0);
+        setSensorDistances(3, 0, 0);
+        
     }
     
     MapCreator::MapCreator(Map p_map)
     {
-      map = p_map;
-      
-      // set Default-values for sensors:
-      setSensorDistances(1, 0, 0);
-      setSensorDistances(2, 0, 0);
-      setSensorDistances(3, 0, 0);
-      
+        map = p_map;
+        
+        // set Default-values for sensors:
+        setSensorDistances(1, 0, 0);
+        setSensorDistances(2, 0, 0);
+        setSensorDistances(3, 0, 0);
+        
     }
     
     MapCreator::~MapCreator()
     {
-      
+        
     }
     
     
     /**
-    * Sets the Distance from Zero
-    * sensor = Number of sensor
-    *  -> 1 = frontsensor           
-    *  -> 2 = sensor on the right
-    *  -> 3 = sensor on the left
-    * x = Distance in x direction
-    * y = Distance in y direction
-    */
-    void MapCreator::setSensorDistances(int sensor, int x, int y)
+     * Sets the Distance from Zero
+     * sensor = Number of sensor
+     *  -> 1 = frontsensor           
+     *  -> 2 = sensor on the right
+     *  -> 3 = sensor on the left
+     * x = Distance in x direction
+     * y = Distance in y direction
+     */
+    void MapCreator::setSensorDistances(int sensor, float x, float y)
     {
-      int tmpX = METER_TO_CENTIMETER(x);
-      int tmpY = METER_TO_CENTIMETER(y);
-      switch(sensor)
-      {
-	case 1: 
-	  sen1.x = tmpX;
-	  sen1.y = tmpY;
-	  break;
-	case 2:
-	  sen2.x = tmpX;
-	  sen2.y = tmpY;
-	  break;
-	case 3:
-	  sen3.x = tmpX;
-	  sen3.y = tmpY;	
-	  break;
-	default:
-	  ROS_WARN("MapCreator: invalid sensor number: %d.", sensor);
-	  break;
-      }
+        int tmpX = (int) METER_TO_CENTIMETER(x);
+        int tmpY = (int) METER_TO_CENTIMETER(y);
+        switch(sensor)
+        {
+            case 1: 
+                sen1.x = tmpX;
+                sen1.y = tmpY;
+                break;
+            case 2:
+                sen2.x = tmpX;
+                sen2.y = tmpY;
+                break;
+            case 3:
+                sen3.x = tmpX;
+                sen3.y = tmpY;	
+                break;
+            default:
+                ROS_WARN("MapCreator: invalid sensor number: %d.", sensor);
+                break;
+        }
     }
-
+    
     void MapCreator::setPosition(RobotPosition p_position)
     {
-      pos.x = METER_TO_CENTIMETER(p_position.point.x) + getXOffset();
-      pos.y = METER_TO_CENTIMETER(p_position.point.y) + getYOffset();
-      pos.theta = p_position.theta * RAD_TO_DEGREE;
+        pos.x = METER_TO_CENTIMETER(p_position.point.x) + getXOffset();
+        pos.y = METER_TO_CENTIMETER(p_position.point.y) + getYOffset();
+        pos.theta = p_position.theta * RAD_TO_DEGREE;
     }
     
     void MapCreator::step(const int p_sensor, const int p_distance)
     {
         calculateObstaclePosition(p_sensor, p_distance);
     }
-
+    
     void MapCreator::calculateObstaclePosition(const int sensor, int measuredDistance)
     {
       
@@ -231,9 +231,9 @@ namespace minotaur {
     
     Map MapCreator::getMap()
     {
-      return map;
+        return map;
     }
-
+    
     void MapCreator::createTextFile(const char *path)
     {
       int x, y;
